Bad Bob! Need to work on MobjCtrl too

+++ gamelib/src/p_local.rs
@@ -4,7 +4,7 @@ pub static ONFLOORZ: i32 = i32::MIN;
 pub static ONCEILINGZ: i32 = i32::MAX;
 
 pub static MAXHEALTH: i32 = 100;
-pub static VIEWHEIGHT: i32 = 41;
+pub static VIEWHEIGHT: f32 = 41.0;
 
 pub static MAXRADIUS: f32 = 32.0;
